backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0
langchain==0.1.5
langchain-core>=0.1
langgraph==0.0.22
langchain-openai==0.0.5
celery==5.3.6
redis==5.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
httpx==0.26.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
aiosqlite==0.19.0  # For in-memory test database
